iommu: Add support for ATS control over sysfs attribute
Add a new per-device IOMMU feature, IOMMU_DEV_FEAT_ATS, to allow the
device drivers to enable or disable the ATS feature for the device.
This could also be called from non-driver context, for example sysfs
interface, if there's no driver binding to the device.

Add the support for users to enable/disable ATS for devices in an iommu
group through the ats sysfs attribute of the iommu group. The users are
required to unbind drivers from the devices before doing this.

The documentation is updated accordingly.

Signed-off-by: Lu Baolu <baolu.lu@linux.intel.com>