iommu/vt-d: Move ATS control to probe/release path
Previously, device ATS (or device TLB) was enabled when a domain
attached to the device, and disabled when the domain detached
from the device. Based on the following considerations, this
refactors this part of logic by moving ATS management to device
probe and release paths:

- After the device is probed, the iommu core usually just switches
  among different domains. It doesn't make much sense to disable
  and then enable ATS in a single callback called from the core.
- For long term, the control of ATS should be managed by the device
  driver. IOMMU driver should avoid controlling it on paths other
  than probe, release and iommu_dev_enable/disable_feature().

On the other hand, moving the ATS control out of the attach/attach
paths implies that the cache operation needs to be changed, because
it ultimately requires to change a bit in a present context entry.

Signed-off-by: Lu Baolu <baolu.lu@linux.intel.com>