iommu/vt-d: Register static iommu devices without locking
The VT-d DMA remapping unit is enumerated by parsing the ACPI DMAR table
during boot, or may be hot added with the PCI root device during runtime.
All units are stored in a list and protected by dmar_global_lock. This
may lead to some potential lockdep splats because the dmar_global_lock
must be held when traversing the list and registering each iommu device.

Introduce a new list of all static drhd units enumerated by parsing the
ACPI DMAR table during boot. Because this list is static throughout the
entire life cycle of the system, there is no need for a lock. Traverse
this list and register the iommu devices in it.

Signed-off-by: Lu Baolu <baolu.lu@linux.intel.com>